Move to the lifecycle of all tests.
